Move Command interface
